+++ src/main/java/racingcar/application/numbergenerator/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.application.numbergenerator;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import racingcar.application.NumberGenerator;
