+++ src/main/java/racingcar/config/AppConfig.java
@@ -1,13 +1,13 @@
 package racingcar.config;
 
-import racingcar.RandomNumberGenerator;
+import racingcar.application.numbergenerator.RandomNumberGenerator;
 import racingcar.application.NameSeparator;
 import racingcar.application.NumberGenerator;
 import racingcar.application.Printer;
 import racingcar.application.Reader;
-import racingcar.nameseparator.OriginSeparator;
-import racingcar.printer.PrintRacingResult;
-import racingcar.reader.ConsoleReader;
+import racingcar.application.nameseparator.OriginSeparator;
+import racingcar.application.printer.PrintRacingResult;
+import racingcar.application.reader.ConsoleReader;
 
 public class AppConfig {
