+++ src/test/java/racingcar/nameseparator/OriginSeparatorTest.java
@@ -3,6 +3,7 @@ package racingcar.nameseparator;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import racingcar.application.nameseparator.OriginSeparator;
 
 class OriginSeparatorTest {
